requirements.txt
jaclang
byllm
pandas
reportlab
python-dotenv
fastapi
uvicorn[standard]
anthropic